# WebAPI/swaggerfile.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "WebAPI",
    "version": "v1"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/departament/createdepartament": {
      "post": {
        "tags": [
          "Departament"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepartamentCreatedDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DepartamentCreatedDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DepartamentCreatedDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/departament/getalldepartament": {
      "get": {
        "tags": [
          "Departament"
        ],
        "parameters": [
          {
            "name": "queryNum",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageNum",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/departament/getbyiddepartament/{id}": {
      "get": {
        "tags": [
          "Departament"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/departament/deletedepartament/{id}": {
      "delete": {
        "tags": [
          "Departament"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/departament/updatedepartament/{id}": {
      "put": {
        "tags": [
          "Departament"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepartamentUpdatedDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DepartamentUpdatedDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DepartamentUpdatedDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Example enter like this => Bearer 'token'",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}